cleaning up trs test cases
